backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==12.0.0
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0
